refactor: move dts emission to utils
